added styling, fixed admin role check

+++ MedicalApp/admin_view.py
@@ -12,17 +12,17 @@ admin_bp = Blueprint('admin', __name__, url_prefix='/admin')
 @admin_bp.route('/')
 @login_required
 def admin_dashboard():
-    #if current_user.access_level != 'ADMIN' or current_user.access_level != 'ADMIN_USER':
-        #return redirect(url_for('home.index'))
+    if current_user.access_level not in ('ADMIN', 'ADMIN_USER'):
+        return redirect(url_for('home.index'))
     user = current_user if current_user.is_authenticated else None
     return render_template('admin_dashboard.html', user=user)
 
 
 @admin_bp.route('/add_user', methods=['GET', 'POST'])
-# @login_required
+@login_required
 def add_user():
-    # if current_user.access_level != 'ADMIN_USER' and current_user.access_level != 'ADMIN':
-        # return redirect(url_for('home.index'))
+    if current_user.access_level not in ('ADMIN', 'ADMIN_USER'):
+        return redirect(url_for('home.index'))
     form = AddUserForm()
     if request.method == 'POST' and form.validate_on_submit():
         email = form.email.data
@@ -46,10 +46,10 @@ def add_user():
 
 
 @admin_bp.route('/delete_user', methods=['GET', 'POST'])
-# @login_required
+@login_required
 def delete_user():
-    # if current_user.access_level != 'ADMIN_USER' and current_user.access_level != 'ADMIN':
-        # return redirect(url_for('home.index'))
+    if current_user.access_level not in ('ADMIN', 'ADMIN_USER'):
+        return redirect(url_for('home.index'))
     form = DeleteUserForm()
     if request.method == 'POST' and form.validate_on_submit():
         email = form.email.data
@@ -64,10 +64,10 @@ def delete_user():
 
 
 @admin_bp.route('/block_user', methods=['GET', 'POST'])
-# @login_required
+@login_required
 def block_user():
-    # if current_user.access_level != 'ADMIN_USER' and current_user.access_level != 'ADMIN':
-        # return redirect(url_for('home.index'))
+    if current_user.access_level not in ('ADMIN', 'ADMIN_USER'):
+        return redirect(url_for('home.index'))
     form = BlockUserForm()
     if request.method == 'POST' and form.validate_on_submit():
         email = form.email.data
@@ -84,10 +84,10 @@ def block_user():
 
 # only ADMIN can do this one
 @admin_bp.route('/change_user_role', methods=['GET', 'POST'])
-# @login_required
+@login_required
 def change_user_role():
-    # if current_user.access_level != 'ADMIN':
-        # return redirect(url_for('home.index'))
+    if current_user.access_level != 'ADMIN':
+        return redirect(url_for('home.index'))
     form = ChangeUserRoleForm()
     if request.method == 'POST' and form.validate_on_submit():
         email = form.email.data
